requirements.txt
fastapi
uvicorn
PyJWT
bcrypt
httpx
passlib
sqlalchemy
python-dotenv
